FaceAI_Appium_Python/requirements.txt
Appium-Python-Client==2.11.1
html-testRunner==1.2.1
selenium==4.15.2 